requirements.txt
aiogram
coc.py
requests
python-dotenv
aiofiles
Levenshtein
